% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbydino.math4753boiii.r
\docType{package}
\name{bbydino.math4753boiii}
\alias{bbydino.math4753boiii}
\title{MATH 4753 - Applied Statistical Methods}
\description{
R Package for MATH 4753 - Applied Statistical Methods
}
\author{
Daniel Ngô \email{bbydino@ou.edu}
}
