% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycontcurve.R
\name{chisqregioncurve}
\alias{chisqregioncurve}
\title{Plot chi-squared distribution and probability region}
\usage{
chisqregioncurve(df, from, to, xlim, ylim, col = "Magenta")
}
\arguments{
\item{df}{df parameter of chisq distribution}

\item{from}{start of probability region}

\item{to}{end of probability region}

\item{xlim}{vector for the x-axis limits}

\item{ylim}{vector for the y-axis limits}

\item{col}{color of the shaded region. default magenta.}
}
\description{
Given df and probability region start and end,
plot chi-squared distribution and probability region.
From MATH 4753 Lab 6.
}
\examples{
chisqregioncurve(df=2, from=2, to=4, xlim=c(-0.5, 20), ylim=c(0, 0.5), col="Pink")

}
