% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycontcurve.R
\name{gammaregioncurve}
\alias{gammaregioncurve}
\title{Plot gamma distribution and probability region}
\usage{
gammaregioncurve(shape, scale, from, to, xlim, col = "Magenta")
}
\arguments{
\item{shape}{shape parameter of gamma distribution}

\item{scale}{scale parameter of gamma distribution}

\item{from}{start of probability region}

\item{to}{end of probability region}

\item{xlim}{vector for the x-axis limits}

\item{col}{color of the shaded region. default magenta.}
}
\description{
Given shape, scale, and probability region start and end,
plot gamma distribution and probability region.
From MATH 4753 Lab 6.
}
\examples{
gammaregioncurve(shape=3, scale=2, from=2, to=5, xlim=c(-.05, 20))
gammaregioncurve(shape=6, scale=3, from=1, to=4, xlim=c(-.05, 20), col="Red")

}
