% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.r
\name{getmode}
\alias{getmode}
\title{Get the Statistical Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{a vector to calculate the statistical mode of.}
}
\value{
the mode of the give vector
}
\description{
Given a vector, return the mode (measure of central tendency). MATH 4753 Ass 1
}
\examples{
getmode(cars$speed)
getmode(c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4))

}
