% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.r
\name{mybin}
\alias{mybin}
\title{Simulate a binomial experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations (default iter=100)}

\item{n}{sample size (default n=10)}

\item{p}{probability of success (default p=0.5)}
}
\value{
table of relative frequencies
}
\description{
Simulates a binomial experiment with given
iteration, sample size, and probability of success.
Outputs a barplot and returns a table of the results.
}
\details{
From: MATH 4753 Lab 5
}
\examples{
sim100 = mybin(iter = 100, n = 10, p = 0.7)
sim1000 = mybin(iter = 1000, n = 10, p = 0.7)

}
