% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.r
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{input data}

\item{fun}{statistic or function to make interval of}

\item{alpha}{alpha value in confidence interval (1-alpha)100 percent confidence}

\item{cx}{text thickness in figure}

\item{...}{params for histogram}
}
\value{
list containing confidence interval, statistic/function, and data
}
\description{
Creates a bootstrap confidence interval for the given statistic
or function.
From: MATH 4753 Lab 9
}
