% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.r
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial - My Central Limit Function}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability of success for each trial}

\item{...}{params for histogram}
}
\description{
Generates n * iter random samples of a binomial distribution
and plots the results.
}
\details{
CLT will work with discrete or continuous distributions.
}
