% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.r
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson - My Central Limit Function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{lambda value of the Poisson random variable}

\item{...}{params for histogram}
}
\description{
Generates n * iter random samples of a Poisson distribution
and plots the results.
}
\details{
CLT will work with discrete or continuous distributions.
}
