% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.r
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform - My Central Limit Theorem}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
list of sample data
}
\description{
Generates n * iter random samples of a uniform distribution,
plots the results, and returns a list of sample values.
}
