% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.r
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{the "DDT" data frame}

\item{SPECIES}{the fish species to plot}
}
\description{
Plot LENGTH vs WEIGHT for a given SPECIES of fish from DDT
From MATH 4753 Project 1.
}
\examples{
myddt(df = ddt, SPECIES = "CCATFISH")
myddt(df = ddt, SPECIES = "SMBUFFALO")

}
