% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.r
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Function for Univariate Likelihoods}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{likelihood function}

\item{x}{input vector}

\item{param}{parameters to the lfun function}

\item{...}{additional plot parameters}
}
\value{
a list containing:
the index of the max likelihood,
the parameter at the index of the max likelihood,
the max likelihood,
the slope
}
\description{
For repeated sampling from same distribution.
}
\details{
From: MATH 4753 Lab 10
}
