% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbinsim.r
\name{mynbinsim}
\alias{mynbinsim}
\title{Simulate a negative binomial experiment}
\usage{
mynbinsim(iter = 100, p = 0.5)
}
\arguments{
\item{iter}{number of iterations (default iter=100)}

\item{p}{probability of success for each trial (default p=0.5)}
}
\value{
table of relative frequencies of probabilities
}
\description{
Simulates a negative binomial experiment with given
iteration, and probability of success for each trial.
Outputs a barplot and returns a table of the results.
}
\details{
From: MATH 4753 Lab 5
}
\examples{
sim100 = mynbinsim(100, 0.5)
sim1000 = mynbinsim(1000, 0.5)

}
