% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycontcurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot normal curve and probability region P(X <= a)}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal distribution}

\item{sigma}{the standard deviation of the normal distribution}

\item{a}{for calculating the area (probability P(X <= a))}
}
\value{
a list containing probability P(X <= a)
}
\description{
Display the curve,
shaded area between the curve and x axis from −∞ to x=a,
and calculate the area (probability, P(X ≤ a)),
which is released to the command-line in a list.
From MATH 4753 Lab 6.
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)

}
