% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.r
\name{myNRML}
\alias{myNRML}
\title{Maximum Likelihood using Newton-Raphson Algorithm}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial x-value for algorithm}

\item{delta}{delta for algorithm, default 0.001}

\item{llik}{likelihood function}

\item{xrange}{x range for algorithm}

\item{parameter}{parameter of interest to the lfun function}
}
\value{
a list containing:
x values from algorithm
y values from algorithm
}
\description{
Maximum Likelihood using Newton-Raphson Algorithm on the derivative
}
\details{
From: MATH 4753 Lab 10
}
