% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycontcurve.R
\name{normregioncurve}
\alias{normregioncurve}
\title{Plot normal curve and probability region}
\usage{
normregioncurve(mean, sd, from, to, col = "Magenta", lower.tail = TRUE)
}
\arguments{
\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}

\item{from}{start of probability region. set to NULL if -inf}

\item{to}{end of probability region. set to NULL if inf}

\item{col}{color of the shaded region. default magenta.}

\item{lower.tail}{if true, get P(X<a). if false, get P(X>a). default to TRUE}
}
\value{
probability of region
}
\description{
Plots a normal distribution curve given
mean, stdev, probability region start, probability region end.
From MATH 4753 Lab 6.
}
\examples{
normregioncurve(mean = 0, sd = 1, from = 2, to = NULL, col = "Cyan", lower.tail = FALSE)
normregioncurve(mean = 4, sd = 2, from = 1, to = 5, col = "Cyan")

}
